+++ chalice/__init__.py
@@ -5,4 +5,4 @@ from chalice.app import (
 )
 
 
-__version__ = '0.2.0'
+__version__ = '0.3.0'